requirements.txt
django
djangorestframework
opencv-python
numpy